Reorganización de de paquetes en: boundary, control, entity. Migración de todos los repositorios a JPA. Arreglado fallo en los avisos con adjuntos.
