+++ src/main/java/es/ucm/fdi/anuncios/business/control/AvisoRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 
-import es.ucm.fdi.anuncios.business.domain.Aviso;
+import es.ucm.fdi.anuncios.business.entity.Aviso;
 
 public interface AvisoRepository extends CrudRepository<Aviso, Long> {
 	public List<Aviso> findByEtiqueta(String etiqueta);
