+++ src/main/java/es/ucm/fdi/anuncios/validation/AvisoValidator.java
@@ -5,7 +5,7 @@ import org.springframework.validation.Errors;
 import org.springframework.validation.ValidationUtils;
 import org.springframework.validation.Validator;
 
-import es.ucm.fdi.anuncios.business.domain.AvisoBuilder;
+import es.ucm.fdi.anuncios.business.entity.AvisoBuilder;
 
 @Component
 public class AvisoValidator implements Validator {
