+++ src/main/java/es/ucm/fdi/espacios/business/boundary/Reservas.java
@@ -1,4 +1,4 @@
-package es.ucm.fdi.espacios.business;
+package es.ucm.fdi.espacios.business.boundary;
 
 import java.util.List;
 
@@ -13,9 +13,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import es.ucm.fdi.espacios.business.control.ReservaRepository;
-import es.ucm.fdi.espacios.business.domain.Espacio;
-import es.ucm.fdi.espacios.business.domain.Reserva;
-import es.ucm.fdi.espacios.business.domain.ReservaBuilder;
+import es.ucm.fdi.espacios.business.entity.Espacio;
+import es.ucm.fdi.espacios.business.entity.Reserva;
+import es.ucm.fdi.espacios.business.entity.ReservaBuilder;
 
 @Service
 @Transactional
