+++ src/main/java/es/ucm/fdi/espacios/business/control/EspacioRepository.java
@@ -2,7 +2,7 @@ package es.ucm.fdi.espacios.business.control;
 
 import org.springframework.data.repository.CrudRepository;
 
-import es.ucm.fdi.espacios.business.domain.Espacio;
+import es.ucm.fdi.espacios.business.entity.Espacio;
 
 public interface EspacioRepository extends CrudRepository<Espacio, Long> {
