+++ src/main/java/es/ucm/fdi/espacios/business/control/ReservaRepository.java
@@ -2,7 +2,7 @@ package es.ucm.fdi.espacios.business.control;
 
 import org.springframework.data.repository.CrudRepository;
 
-import es.ucm.fdi.espacios.business.domain.Reserva;
+import es.ucm.fdi.espacios.business.entity.Reserva;
 
 public interface ReservaRepository  extends CrudRepository<Reserva, Long> {
