+++ src/main/java/es/ucm/fdi/espacios/business/entity/Reserva.java
@@ -1,21 +1,33 @@
-package es.ucm.fdi.espacios.business.domain;
+package es.ucm.fdi.espacios.business.entity;
 
+import javax.persistence.CascadeType;
+import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 
 import org.hibernate.annotations.Type;
 import org.joda.time.DateTime;
-import org.springframework.beans.BeanUtils;
 import org.springframework.format.annotation.DateTimeFormat;
 
-public class ReservaBuilder {
-
+@Entity
+public class Reserva {
+	@Id
+	@GeneratedValue(strategy = GenerationType.AUTO)
 	private Long id;
-
+	
+	@ManyToOne(optional=true)
+	@JoinColumn(name="ESPACIO_ID")
+	private Espacio espacio;
+	
 	private String titular;
-
+	
 	//Fechas
-
+	
 	@Type(type = "org.jadira.usertype.dateandtime.joda.PersistentDateTime")
 	@DateTimeFormat(pattern = "yyyy/mm/dd HH:mm")
 	private DateTime fechaCreacion;
@@ -27,22 +39,18 @@ public class ReservaBuilder {
 	@Type(type = "org.jadira.usertype.dateandtime.joda.PersistentDateTime")
 	@DateTimeFormat(pattern = "yyyy/mm/dd HH:mm")
 	private DateTime fechaFin;
-
+	
 	private String aclaraciones;
-
+	
 	@Enumerated(EnumType.ORDINAL)
 	private TipoEspacioEnum tipoEspacioEnum;
-	
-	private Long id_espacio;
-	
-	public ReservaBuilder(){
+
+	public Reserva() {
 		
 	}
 	
-	public Reserva build(){
-		Reserva nuevo = new Reserva(this.id);
-		BeanUtils.copyProperties(this, nuevo);
-		return nuevo;
+	public Reserva(Long id) {
+		this.id = id;
 	}
 
 	public Long getId() {
@@ -101,20 +109,16 @@ public class ReservaBuilder {
 		this.tipoEspacioEnum = tipoEspacioEnum;
 	}
 
-	/**
-	 * @return the id_espacio
-	 */
-	public Long getId_espacio() {
-		return id_espacio;
+	public Espacio getEspacio() {
+		return espacio;
 	}
 
-	/**
-	 * @param id_espacio the id_espacio to set
-	 */
-	public void setId_espacio(Long id_espacio) {
-		this.id_espacio = id_espacio;
+	public void setEspacio(Espacio espacio) {
+		this.espacio = espacio;
 	}
 	
 	
 
+	
+	
 }
