--- src/main/resources/hibernate.cfg.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-configuration PUBLIC
-        "-//Hibernate/Hibernate Configuration DTD 3.0//EN"
-        "http://www.hibernate.org/dtd/hibernate-configuration-3.0.dtd
-">
-<hibernate-configuration>
-	<session-factory>
-		<!-- Database connection settings -->
-		<property name="connection.driver_class">
-			org.hsqldb.jdbc.JDBCDriver
-		</property>
-		<property name="connection.url">
-			jdbc:hsqldb:db2;shutdown=true
-		</property>
-		<property name="connection.username">sa</property>
-		<property name="connection.password" />
-		<property name="dialect">org.hibernate.dialect.HSQLDialect</property>
-		<!-- Echo all executed SQL to stdout -->
-		<property name="show_sql">true</property>
-		<!-- Drop and re-create the database schema on startup -->
-		<!-- <property name="hbm2ddl.auto">create-drop</property> -->
-		<!-- <property name="hbm2ddl.auto">create</property>
-		<property name="hbm2ddl.auto">validate</property>	 -->	
-		<property name="hbm2ddl.auto">update</property>
-		<mapping class="es.ucm.fdi.anuncios.business.domain.Aviso"/>
-		<mapping class="es.ucm.fdi.store.business.entity.StorageObject" />
-	</session-factory>
-
-</hibernate-configuration>
