--- src/main/webapp/WEB-INF/spring/webcontext/espacios-context.xml
@@ -1,154 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:p="http://www.springframework.org/schema/p"
-	xmlns:context="http://www.springframework.org/schema/context"	
-	xmlns:mvc="http://www.springframework.org/schema/mvc" xmlns:tx="http://www.springframework.org/schema/tx"
-	xmlns:util="http://www.springframework.org/schema/util" xmlns:data="http://www.springframework.org/schema/data/jpa"
-	xmlns:jpa="http://www.springframework.org/schema/data/jpa"
-	xmlns:jdbc="http://www.springframework.org/schema/jdbc"
-
-	xsi:schemaLocation="http://www.springframework.org/schema/beans
-	http://www.springframework.org/schema/beans/spring-beans.xsd
-	http://www.springframework.org/schema/tx
-	http://www.springframework.org/schema/tx/spring-tx.xsd
-	http://www.springframework.org/schema/util
-	http://www.springframework.org/schema/util/spring-util.xsd
-	http://www.springframework.org/schema/context
-	http://www.springframework.org/schema/context/spring-context-4.0.xsd
-	http://www.springframework.org/schema/mvc
-	http://www.springframework.org/schema/mvc/spring-mvc-4.0.xsd
-	http://www.springframework.org/schema/data/jpa 
-	http://www.springframework.org/schema/data/jpa/spring-jpa.xsd">
-
-	<context:component-scan base-package="es.ucm.fdi.espacios.business" />
-
-
-	<jpa:repositories base-package="es.ucm.fdi.espacios.business.domain.repository" />
-	
-	<bean class="org.springframework.orm.jpa.JpaTransactionManager">
-		<property name="entityManagerFactory" ref="entityManagerFactory" />
-	</bean>
-	<bean id="entityManagerFactory"
-		class="org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean">
-		<property name="dataSource" ref="espaciosDataSource" />
-		<property name="packagesToScan" value="es.ucm.fdi.espacios.business.domain" />
-		<property name="jpaVendorAdapter">
-			<bean class="org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter">
-				<property name="database" value="HSQL" />
-				<property name="generateDdl" value="true" />
-			</bean>
-		</property>
-	</bean>
-	
-
-
-	<!-- ESPACIOS -->
-
-	<!-- <data:repositories base-package="es.ucm.fdi.reservas.business.domain.repository" 
-		/> -->
-	<!-- entity-manager-factory-ref="espaciosEntityManagerFactory" transaction-manager-ref="espaciosTransactionManager" 
-		/> -->
-
-	<!-- <bean id="espaciosEntityManagerFactory" -->
-	<!-- class="org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean"> -->
-	<!-- <property name="dataSource" ref="espaciosDataSource" /> -->
-	<!-- <property name="loadTimeWeaver"> -->
-	<!-- <bean -->
-	<!-- class="org.springframework.instrument.classloading.InstrumentationLoadTimeWeaver" 
-		/> -->
-	<!-- </property> -->
-	<!-- <property name="packagesToScan"> -->
-	<!-- <array> -->
-	<!-- <value>es.ucm.fdi.espacios.business.domain</value> -->
-	<!-- </array> -->
-	<!-- </property> -->
-	<!-- <property name="jpaVendorAdapter" ref="espaciosJpaVendorAdapter" /> -->
-	<!-- <property name="jpaProperties"> -->
-	<!-- <props> -->
-	<!-- <prop key="hibernate.dialect">org.hibernate.dialect.HSQLDialect</prop> -->
-	<!-- <prop key="hibernate.max_fetch_depth">3</prop> -->
-	<!-- <prop key="hibernate.jdbc.fetch_size">50</prop> -->
-	<!-- <prop key="hibernate.jdbc.batch_size">10</prop> -->
-	<!-- <prop key="hibernate.show_sql">true</prop> -->
-	<!-- <prop key="hibernate.hbm2ddl.auto">create-drop</prop> -->
-	<!-- </props> -->
-	<!-- </property> -->
-	<!-- </bean> -->
-
-	<!-- <bean id="espaciosJpaVendorAdapter" -->
-	<!-- class="org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter" 
-		/> -->
-
-	<bean id="espaciosDataSource" class="org.apache.commons.dbcp.BasicDataSource"
-	destroy-method="close">
-	<property name="driverClassName" value="org.hsqldb.jdbcDriver" />
-	<property name="url" value="jdbc:hsqldb:espacios;shutdown=true" />
-	<property name="username" value="sa" />
-	<property name="password" value="" />
-	</bean>
-
-	<!-- <bean id="espaciosTransactionManager" class="org.springframework.orm.jpa.JpaTransactionManager" -->
-	<!-- p:entityManagerFactory-ref="espaciosEntityManagerFactory" /> -->
-
-
-
-
-	<!-- RESERVAS -->
-
-	<!-- <data:repositories base-package="es.ucm.fdi.espacios.business.domain.repository" -->
-	<!-- entity-manager-factory-ref="reservasEntityManagerFactory" -->
-	<!-- transaction-manager-ref="reservasTransactionManager" /> -->
-
-	<!-- <bean id="reservasEntityManagerFactory" -->
-	<!-- class="org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean"> -->
-	<!-- <property name="dataSource" ref="reservasDataSource" /> -->
-	<!-- <property name="loadTimeWeaver"> -->
-	<!-- <bean -->
-	<!-- class="org.springframework.instrument.classloading.InstrumentationLoadTimeWeaver" 
-		/> -->
-	<!-- </property> -->
-	<!-- <property name="packagesToScan"> -->
-	<!-- <array> -->
-	<!-- <value>es.ucm.fdi.espacios.business.domain</value> -->
-	<!-- </array> -->
-	<!-- </property> -->
-	<!-- <property name="jpaVendorAdapter" ref="reservasJpaVendorAdapter" /> -->
-	<!-- <property name="jpaProperties"> -->
-	<!-- <props> -->
-	<!-- <prop key="hibernate.dialect">org.hibernate.dialect.HSQLDialect</prop> -->
-	<!-- <prop key="hibernate.max_fetch_depth">3</prop> -->
-	<!-- <prop key="hibernate.jdbc.fetch_size">50</prop> -->
-	<!-- <prop key="hibernate.jdbc.batch_size">10</prop> -->
-	<!-- <prop key="hibernate.show_sql">true</prop> -->
-	<!-- <prop key="hibernate.hbm2ddl.auto">create-drop</prop> -->
-	<!-- </props> -->
-	<!-- </property> -->
-	<!-- </bean> -->
-
-	<!-- <bean id="reservasJpaVendorAdapter" -->
-	<!-- class="org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter" 
-		/> -->
-
-	<!-- <bean id="reservasDataSource" class="org.apache.commons.dbcp.BasicDataSource" -->
-	<!-- destroy-method="close"> -->
-	<!-- <property name="driverClassName" value="org.hsqldb.jdbcDriver" /> -->
-	<!-- <property name="url" value="jdbc:hsqldb:reservas;shutdown=true" /> -->
-	<!-- <property name="username" value="sa" /> -->
-	<!-- <property name="password" value="" /> -->
-	<!-- </bean> -->
-
-	<!-- <tx:annotation-driven transaction-manager="espaciosTransactionManager" 
-		/> -->
-
-	<!-- <bean id="reservasTransactionManager" class="org.springframework.orm.jpa.JpaTransactionManager" 
-		p:entityManagerFactory-ref="reservasEntityManagerFactory" /> -->
-
-	<!-- -->
-	<!-- Spring MVC configuration -->
-	<!-- -->
-
-	<!-- Define Spring MVC configuration using java annotations -->
-	<mvc:annotation-driven />
-
-	<context:component-scan base-package="es.ucm.fdi.espacios.web" />
-</beans>
