--- src/main/webapp/WEB-INF/spring/webcontext/storage-context.xml
@@ -1,84 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xmlns:p="http://www.springframework.org/schema/p"
-	xmlns:context="http://www.springframework.org/schema/context"
-	xmlns:mvc="http://www.springframework.org/schema/mvc"
-	xmlns:tx="http://www.springframework.org/schema/tx"
-	xmlns:util="http://www.springframework.org/schema/util"
-	xmlns:data="http://www.springframework.org/schema/data/jpa"
-
-	xsi:schemaLocation="http://www.springframework.org/schema/beans
-	http://www.springframework.org/schema/beans/spring-beans.xsd
-	http://www.springframework.org/schema/tx
-	http://www.springframework.org/schema/tx/spring-tx.xsd
-	http://www.springframework.org/schema/util
-	http://www.springframework.org/schema/util/spring-util.xsd
-	http://www.springframework.org/schema/context
-	http://www.springframework.org/schema/context/spring-context-4.0.xsd
-	http://www.springframework.org/schema/mvc
-	http://www.springframework.org/schema/mvc/spring-mvc-4.0.xsd
-	http://www.springframework.org/schema/data/jpa
-	http://www.springframework.org/schema/data/jpa/spring-jpa.xsd">
-
-	<context:component-scan base-package="es.ucm.fdi.storage.business.boundary" />
-
-	<util:properties id="storagePrefs">
-		<prop key="rootPath">WEB-INF/storage</prop>
-		<prop key="serviceURL">http://localhost:8080/anuncios/storage/</prop>
-	</util:properties>
-
-	<data:repositories base-package="es.ucm.fdi.storage.business.control"
-		entity-manager-factory-ref="storageEntityManagerFactory"
-		transaction-manager-ref="storageTransactionManager" />
-
-	<bean id="storageEntityManagerFactory"
-		class="org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean">
-		<property name="dataSource" ref="storageDataSource" />
-		<property name="loadTimeWeaver">
-			<bean
-				class="org.springframework.instrument.classloading.InstrumentationLoadTimeWeaver" />
-		</property>
-		<property name="packagesToScan">
-			<array>
-				<value>es.ucm.fdi.storage.business.entity</value>
-			</array>
-		</property>
-		<property name="jpaVendorAdapter" ref="storageJpaVendorAdapter" />
-		<property name="jpaProperties">
-			<props>
-				<prop key="hibernate.dialect">org.hibernate.dialect.HSQLDialect</prop>
-				<prop key="hibernate.max_fetch_depth">3</prop>
-				<prop key="hibernate.jdbc.fetch_size">50</prop>
-				<prop key="hibernate.jdbc.batch_size">10</prop>
-				<prop key="hibernate.show_sql">true</prop>
-				<prop key="hibernate.hbm2ddl.auto">create-drop</prop>
-			</props>
-		</property>
-	</bean>
-
-	<bean id="storageJpaVendorAdapter" class="org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter" />
-
-	<bean id="storageDataSource" class="org.apache.commons.dbcp.BasicDataSource"
-		destroy-method="close">
-		<property name="driverClassName" value="org.hsqldb.jdbcDriver" />
-		<property name="url" value="jdbc:hsqldb:storage;shutdown=true" />
-		<property name="username" value="sa" />
-		<property name="password" value="" />
-	</bean>
-
-	<tx:annotation-driven transaction-manager="storageTransactionManager"/>
-	
-	<bean id="storageTransactionManager" class="org.springframework.orm.jpa.JpaTransactionManager"
-		p:entityManagerFactory-ref="storageEntityManagerFactory" />
-
-
-	<!-- -->
-	<!-- Spring MVC configuration -->
-	<!-- -->
-
-	<!-- Define Spring MVC configuration using java annotations -->
-	<mvc:annotation-driven />
-
-	<context:component-scan base-package="es.ucm.fdi.storage.web" />
-</beans>
