+++ src/main/webapp/WEB-INF/web.xml
@@ -164,7 +164,7 @@
 		<servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
 		<init-param>
 			<param-name>contextConfigLocation</param-name>
-			<param-value>/WEB-INF/spring/webcontext/portal-context.xml, /WEB-INF/spring/webcontext/storage-context.xml, /WEB-INF/spring/webcontext/espacios-context.xml</param-value>
+			<param-value>/WEB-INF/spring/webcontext/portal-context.xml</param-value>
 		</init-param>
 		<load-on-startup>1</load-on-startup>
 	</servlet>
